+++ queue/export_test.go
@@ -1,0 +1,3 @@
+package queue
+
+func NewMessageIDForTest() MessageID { return newMessageID() }